Add files via upload
These are the some of the analysis performed in this project.
Q. 1) Convert the Datatype of 'Date' column to Date-Time format.

Q. 2) 
    *** Add a new column ''year'' in the dataframe, which contains years only.
    *** Add a new column ''month'' as 2nd column in the dataframe, which contains month only.

Q. 3) Remove the columns 'year' and 'month' from the dataframe.

Q. 4) Show all the records where 'No. of Crimes' is 0. And, how many such records are there ?

Q. 5) What is the maximum & minimum 'average_price' per year in england ?

Q. 6) What is the Maximum & Minimum No. of Crimes recorded per area ?

Q. 7) Show the total count of records of each area, where average price is less than 100000